backend/requirements.txt
Django==4.1.7
django-cors-headers==3.14.0
django-import-export==3.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2==2.9.5
PyJWT==2.6.0